Remove js cookie getters from content-script. Add separate js cookie inject into all frames.
